Crate xmpp_events

Structs

Auth

Bind

CloseStream

GenericIq

GenericMessage

Message

OpenStream

Ping

Presence

ProceedTls

StartTls

StreamFeatures

SuccessTls

Enums

Event
Define an Event between a client and a server (or a server and another server)

EventType

Features
Define possible features found in a stream:features

IqEvent
Define a sub part of an Event, an IqEvent is the representation of a Stanza which is an IQ.

IqType

NonStanzaEvent
Define a sub part of an Event, a NonStanzaEvent is the representation of an XML Stream event. It's used by the system to deal with the communication between entities over a network.

PresenceType

StanzaEvent
Define a sub part of an Event, a StanzaEvent is the representation of a Stanza which can be a Presence, an IQ or a Message.

Traits

FromGeneric

FromXmlElement
FromXmlElement is used to transform an Element to an object

ToEvent
ToEvent is used to transform an object to an XMPP Event

ToXmlElement
ToXmlElement is used to transform any struct to an XML Element