Enum xmpp_events::Event

pub enum Event {
    NonStanza(Box<NonStanzaEvent>),
    Stanza(Box<StanzaEvent>),
}

Define an Event between a client and a server (or a server and another server)

An Event has many types. First, it can be an a NonStanza or a Stanza.

A NonStanza is an XML Stream, an XML stream is a container for the exchange of XML elements between any two entities over a network.

A Stanza is an XML Stanza, an XML stanza is a discrete semantic unit of structured information that is sent from one entity to another over an XML stream.

a ping Event is define like this:

Event::Stanza(Box<StanzaEvent::IqEvent(Box<IqEvent::PingEvent(Ping)>)>)

It complicated but every Event can be instanciate in a more clearer way:

extern crate xmpp_jid;
extern crate xmpp_events;

use xmpp_events::{ToEvent, Event, IqEvent, Message, Ping,  StanzaEvent};
use xmpp_jid::Jid;
use std::str::FromStr;
 
fn main() {
    let from = Jid::from_str("from_jid").unwrap();
    let to = Jid::from_str("to_jid").unwrap();
 
    // Create a Ping event (which is a struct with some Derive and implementing ToEvent)
    let ping = Ping::new(from, to);
 
    // ToEvent will transform a simple struct into a full path Stanza/NonStanza
    let x = ping.to_event(); // produce the full event path
 
 
    // We can check that the produced event are exactly what we want,
    // A Stanza which is an IqEvent (can be both Error,Result,Get,Set)
    // Of the predefined type: PingEvent
    match x {
        Event::Stanza(stanza) => match *stanza {
            StanzaEvent::IqEvent(iq) => match *iq {
                IqEvent::PingEvent(ref ping) => { 
                    // Deal with ping
                },
                _ => {}
            },
            _ => {}
        },
        _ => {}
    }
}

Variants
NonStanza(Box<NonStanzaEvent>)

Represent an event which is an XML Stream

Stanza(Box<StanzaEvent>)

Represent an event which is an XML Stanza

Methods

impl Event

fn is_message(&self) -> bool

Test if the current Event is a Message event type

Examples

extern crate xmpp_jid;
extern crate xmpp_events;

use xmpp_events::{ToEvent, Message};
use xmpp_jid::Jid;
use std::str::FromStr;

fn main() {
    let m = Message::new(Jid::from_str("someJId").unwrap(), "heyy");

    let event = m.to_event();

    assert!(event.is_message());
}

fn is_presence(&self) -> bool

Test if the current Event is a Presence event type

Examples

extern crate xmpp_jid;
extern crate xmpp_events;

use xmpp_events::{ToEvent, Message};
use xmpp_jid::Jid;
use std::str::FromStr;

fn main() {
    let m = Message::new(Jid::from_str("someJId").unwrap(), "heyy");

    let event = m.to_event();

    assert!(!event.is_presence());
}

fn is_iq(&self) -> bool

Test if the current Event is an IQ event type

Examples

extern crate xmpp_jid;
extern crate xmpp_events;

use xmpp_events::{ToEvent, Message};
use xmpp_jid::Jid;
use std::str::FromStr;

fn main() {
    let m = Message::new(Jid::from_str("someJId").unwrap(), "heyy");

    let event = m.to_event();

    assert!(!event.is_iq());
}

fn is_non_stanza(&self) -> bool

Test if the current Event is a NonStanza event type

Examples

extern crate xmpp_jid;
extern crate xmpp_events;

use xmpp_events::{ToEvent, Message};
use xmpp_jid::Jid;
use std::str::FromStr;

fn main() {
    let m = Message::new(Jid::from_str("someJId").unwrap(), "heyy");

    let event = m.to_event();

    assert!(!event.is_non_stanza());
}