Enum xmpp_events::StanzaEvent

pub enum StanzaEvent {
    PresenceEvent(Presence),
    IqEvent(Box<IqEvent>),
    IqRequestEvent(Box<IqEvent>),
    IqResponseEvent(Box<IqEvent>),
    MessageEvent(Box<GenericMessage>),
}

Define a sub part of an Event, a StanzaEvent is the representation of a Stanza which can be a Presence, an IQ or a Message.

Variants
PresenceEvent(Presence)IqEvent(Box<IqEvent>)IqRequestEvent(Box<IqEvent>)IqResponseEvent(Box<IqEvent>)MessageEvent(Box<GenericMessage>)